requirements.txt
flask==2.3.3
torch==2.3.0
torchvision==0.18.0
scikit-learn>=1.3.0
adversarial-robustness-toolbox>=1.18.0 
python-dotenv==1.0.0
google-generativeai==0.1.0
markupsafe==2.1.3
matplotlib==3.8.2
pillow==10.2.0